delete deprecated components + story move new components to NotificationsList directory
