+++ packages/gamut/src/NotificationList/EmptyNotification/emptyNotificationContents.tsx
