--- packages/gamut/src/NotificationList/styles/Notification.module.scss
@@ -1,43 +1,0 @@
-@import "~@codecademy/gamut-styles/utils";
-
-.notification {
-  background: none;
-  width: 100%;
-  padding: 1rem 1.5rem;
-  display: flex;
-  align-items: flex-start;
-  font-size: 0.875rem;
-  color: $color-black;
-  border: none;
-  border-bottom: 1px solid $color-gray-200;
-  text-align: left;
-
-  &.unread {
-    background-color: rgba($color-blue-500, 0.1);
-
-    .time {
-      color: $color-gray-700;
-    }
-  }
-
-  &:last-child {
-    border-bottom: 0;
-  }
-
-  &:hover {
-    text-decoration: none;
-  }
-}
-
-.body {
-  width: 100%;
-}
-
-.text {
-  max-height: 4rem;
-  overflow: hidden;
-}
-
-.time {
-  color: $color-gray-500;
-}
