--- packages/gamut/src/NotificationList/styles/NotificationIcon.module.scss
@@ -1,25 +1,0 @@
-.iconContainer {
-  height: 3rem;
-  width: 3rem;
-  margin-right: 1.5rem;
-}
-
-.iconComponentContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50%;
-  height: 3rem;
-  width: 3rem;
-}
-
-.icon {
-  height: 3rem;
-  width: 3rem;
-  border-radius: 50%;
-}
-
-.iconImage {
-  height: 2.5rem;
-  width: 2.5rem;
-}
