--- packages/gamut/src/NotificationList/styles/index.module.scss
@@ -1,21 +1,0 @@
-@import "~@codecademy/gamut-styles/utils";
-
-.notificationsContainer {
-  background-color: $color-white;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  font-size: 1rem;
-  color: $color-gray-800;
-}
-
-.emptyContainer {
-  padding: 3rem;
-}
-
-.emptyText {
-  background: none;
-  border: none;
-  text-align: center;
-}
