+++ packages/gamut/src/index.tsx
@@ -37,10 +37,7 @@ export * from './Menus/SideMenu';
 export * from './Dialog';
 export * from './Modal';
 export * from './NotificationList';
-export * from './NotificationList/NotificationIcon';
-export * from './NotificationList/NotificationItem';
-export * from './NotificationListNew';
-export * from './NotificationListNew/NotificationItemNew';
+export * from './NotificationList/NotificationItemNew';
 export * from './Overlay';
 export * from './Pattern';
 export * from './ProgressBar';
